Reviewing  tests to setup project for future developments.
https://flask.palletsprojects.com/en/3.0.x/tutorial/tests/

'by test on 2018-01-01' not found in response.
Content tests are now wrong since text in the app has been changed to be in french.

Tests are good for now. However the update button showing on articles if the user logged in is the author of the article is gone at the moment.

The next step is thus to restore this functionality before getting into the implementation of new functionalities.