Working on file upload for posts. User should be able to upload images to their posts. The image is of course to be displayed on the post in the blog.
At the moment the form is changed in post creation to add an image field.

The field doesn't update when the image is put in the form, making hard know if the selection went correctly. Work needs to be done to update dynamically the field value.

The file is correctly received in the flask backend. The file is written to disk for now.

We can try to reread the file from disk to display it in the post.