Working on file upload for posts. User should be able to upload images to their posts. The image is of course to be displayed on the post in the blog.
Testing flask extension WTF that implements standards forms.

Added a new temporary page /wtf to conduct the testing.

The form as is is functional and when a file is uploaded the name of the file is automatically displayed which is some kind of feedback.

Writing a WTF forms does require way less lines of code to achieves the same if not more functionality than trying to use html forms and http requests.